Feat:completei os controler de conta e finalizaei a classse pagamento falta implementar o lancamento de itens de conta

+++ backend/src/main/java/ufpi/ivangusthavo/mybar/controller/ClienteController.java
@@ -38,7 +38,7 @@ public class ClienteController {
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<?> deletarCliente(@PathVariable Integer id)
+    public ResponseEntity<?> deletarCliente(@PathVariable Long id)
     {
 
         dao.deleteById(id);
