+++ backend/src/main/java/ufpi/ivangusthavo/mybar/repository/InterfaceCliente.java
@@ -4,5 +4,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ufpi.ivangusthavo.mybar.model.Cliente;
 
-public interface InterfaceCliente extends JpaRepository<Cliente, Integer> {
+public interface InterfaceCliente extends JpaRepository<Cliente, Long> {
 }
